PDS API 52: add instruments, investigations, and targets (#8)
* make a docker system

* more docker

* strings work

targets and investigations within the elastic search output are strings not lists. These work. The instrument is a list and the JSON property helper fails.

* arrays work

Made all 3 be arrays and allow single item as array. If in the future there are mixed arrays and singles of instruments, investigations, and targets, this code should work just fine.

* add host

The example shows instrument_host as well as instrument making up the observing_system_components so added it.

* these are the properties needed to run service in docker

* fix review requests

* replace the necessary part

Over enthusiastic cut so putting back the good bits.

* move methods about

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>