meet acceptance criteria
First, added verify/issue_4.py. This requires requests and python3 but verifies that the URL returns with the expected status code. During the making of this script, discovered that some of the URLs in the ticket where collections/bundles. Reduced these to just bundles and they worked.

One problem was the SearchHits iterator was being invoked when there were not items in the iterator. Added a check to verify that there were more and returned an iterator to an empty list when not.

With the iterator fixed, looked at the products returned to make sure it was not an empty list. When it is, return the 404 page.

Lastly, when given a lidvid without a version finding the latest was failing because the list was empty. Added a check that if the list is empty to return the original lid allowing the already existing checks to take place and return the 404 instead of a 500 from an index out of bounds.