update lexer/parser
"a eq b" does not work becasue b is/was parsed to a FIELD token. Because of this, had to modify the lexer to allow a FIELD on both left and right side of the operator. The unit test that showed the parsing failure of "a eq b" now passes.