pds-api-56: crawl the hierarchical tree (#29)
* start with the acceptance criteria being checked

* cosmetic changes

* update the API

The POM needs to request the correct version of the api JAR file.

Update the 3 API interfaces with the new methods. Currently they are blank place holders and the appropriate branches will be created to fill in each of the bodies. The branches will be created from here then merged back.

* basics for products of a bundle

* now get products of a bundle. works and tested

* prepare for finding owners

* implement bundle from collection

Basically use the search capabilities to find where the collection lid shows up in the ref_lid_collection variable. Pretty straight forward.

* editorial

* fix what should not have changed

* find bundles of a product

Pretty straight foward finding all collections that hold a this product then all bundles that reference those collections.

* collections that reference product

* fix whitespacing

* get actual collection

* do negative checks

If the lidvid is wrong, make sure that nothing is returned. This was added because it was found that the double search of /products/x/bundles returned multiple responses instead of 0. Because the second search had no criteria it returned the entire database. Added code to prevent this then added negative checks for all URLs.

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>