make ES do the work
Made a common method that fetches just the desired fields of a list of lidvids. The idea is that the list of lidvids is the pagination and then ES does the work of stripping it down.

To test the idea, updated the /bundles/<lidvid>/products routine to use the common method.