Issue 13.1 (#45)
* bundles more efficent

While bringing the patch across from branch issue_13, refactored it a bit and now have a re-usable bit of code with elastic search stuff in the ElasticSearchUtil class and some PDS product stuff in buisness objects. There is a mix of stuff that uses the elastic seaarch stuff and pushes it to buisness objects that was left in the controllers. May refactor some that as it is fairly repitive too.

* refactored some more

Moved stuff common to bundles, collections, and products to the base class.

* update collections controller

* update product

* merge of main repair

* make all fo issue_56 checks work

The code changes are basically for the walking forward/backward in the product tree more efficiently so used verify/issue_56.py to show that they still function correctly.

* also collect default fields

* request maximum data

* request maximum data

* request maximum data

* request maximum data

* revert and find another way

* updated the bundle endpoint

Both collections and products from a bundle lidvid should not work through all of the available matching data.

* second rework

Removed ElasticSearchUitl.collate() and updated much of the looping so that it would find all items. It was the bulk of the work. When looking child to parent, the start and limit are used as part of the search request for which collate worked well and thus put back.

* ++ operator behaves oddly

* fix auto checks

* gamma debug messages

* too much

Seems that asking for too much from ES causes it to really slow down. For that reason, changed some of the outer paging stuff back to a fixed default 10 items.

* making a lidvid is expensive??

* have to jump forward at times

* delay is aback to large gets from ES

* more auto check fixes

* more timing data

* test no lid to lidvid safety

* option 1

* requested updates

* add comments

* remove extra use of limit

* remove extra use of limit

* get full list of products if ready

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>