pds-api 68 : Add hits and took (#61)
* keyword removed but not

Found that the API had removed the keyword in the arguments but never updated any of the code. Not sure on the correct fix, so just half removed it in this commit. If it really should be removed, then that is another ticket to finish the job.

* add hits and tooks

Used the eclipse search to find all creations of the Summary then added hits and tooks to all of them.

* Revert "keyword removed but not"

This reverts commit 9090b5615a1db1691ed40828dcf62437bdeb5fca.

* Removed the '::' checks

The call to convert lid to lidvid is more comprehensive than the if states which in turn will break the desire that non-existent lidvid return a 404.

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>
Co-authored-by: Jordan Padams <jordan.h.padams@jpl.nasa.gov>
Co-authored-by: Jordan Padams <33492486+jordanpadams@users.noreply.github.com>