fix what is returned when fields is blanks
Accidentally made it so only the minimal set was returned if fields are left blank. Added a min/max list of fields where min must be a subset of max and empty is a full set without having to define all that may be in elastic search. This min/max approach fixes the blank fields and adds constraints so that pds+ formats will only  every reeturn the allowable keys and nothing else. These are simple to change in the future as well by modifying just the min/max arrays.