amazing amount of cleanup
Turns out there is a lot of duplicate code that wanted to do the same thing but had different behaviors and end results. This really made testing and comparing odd because it would fail for one user but not another mostly because of endpoints used when there should have been no difference.  This change cleans most of that up with respect to getting data and using the fields to include and those to exclude. The blob has successfully been removed once again.