changes blob handling
First, there are now two blob possibilities and they are encoded. Added the second blob type and adjusted the searching to handle either or both or neither.

The Pds4Prduct JSON serializer did the decoding of the blob. Moved it from the single serializer to the Pds4ProductFactory where all of the creation happens. Now the blob is decoded for PDS4 only but for any PDS4 output format.