add the functional error messages
First added a fectory to build the model.ErrorMessage from the original request and the exception thrown.

Second, modified all the controllers to create the ResponseEntity with the model.ErrorMessage as well as the error message. Might need/want to the http error code and name of the Java exception (class name) to the output.

Third, built all of the serializers for the various output types. Rather than one serializer for each type, squashed some of the types down like all JSON are the same and all XML are the same.

Fourth and lastly, update WebMVCConfig with all of the new serializers.