first step toward opensearch
Refactor two of the packages to remove the elasticsearch moniker. Not a big deal, but will make it clearer. First elasticsearch.business is renamed to business because there is little to know specific logic to elasticsearch in it. It is more translating domain specific content to webby content. Second, renamed elasticsearch to just search. The bulk of the package is really a portal (model?) to the search engine in the back. Therefore called it just search since it could support both open and elastic if desired. Doing so would require an explicit set of interfaces, but the code is separated enough it functions as a portal or facade pattern.