issue 106: consistent behavior and documentation (#112)
* update swagger

Make the swagger documentation match the current behavior. Made them all have the same parameters based on their output being singular or plural and then what is controllable with respect to that output.

Much of the Java controllers had to be updated to match the new parameters etc. Those also have all been updated.

* cleaned up swagger docs

Went through the swagger YAML and cleaned up the descriptions for the endpoints. Cleaned up the processing to begin conforming to swagger YAML.

* forgot the new class

* back to uniform flow (nearly)

Have all endpoints go through same search builder to help with improving performance (change one place) and control (tracking status). Found that some may not be using the builder so will need to fix that.

* major restructure to have top->down control

* all restructured and dead code removed

* general fixes

* fix all things kvp

* reduce all lidvid lookups to one

* more deadcode removal

* more deadcode removal

* everything back to working but all

* all endpoints work again

* all endpoints work again

* add some so-so documentation

* clean up

* changed home controller

* updates

* additional cleanup after merge

* fixed some sona-type

* added tracking meta archive status

* changed match to term and should to must

* more should/must clean up

* more should/must clean up

* some more rework

* doc update

* sonatype fixes

* finish cleaun up between merge and other refactoring

* some help in checking the status of the code

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>