general bug fixes (a collection) (#166)
* Issue 131: increase adaptability (#142)

* endpoints

Had to change the planned endpoints to /gid/{group name} to /pid/{lidvid}. Also had to add /gid and /pid as prefixes for the referencing endpoints as well. The point really is that had to add the prefixes is because swagger cannot tell /{group name} from /{lidvid} even though the group names are a well defined enum. The prefixes allow swagger to identify the endpoints as unique. While the actual characters gid and pid are not required, whatever is chosen must be unique and different.

* reshape the architecture/design

The base architecture is MVC, but the code does not directly represent it. Moved business to model and serializer to view to match the architecture. Reduced the multiple controllers to one and use a facade/plug-n-play pattern (ReferencingLogic) to manage the polymorphism needed  for each of the ancestral paths to find information.

Added a Java enum to do the mapping among the group names and the product class in the data itself. The enum holds the ReferencingLogic allowing either side (controller or model) to work with the given inputs and collect the desired output.

Most of the code in the DAO objects moved to ReferencingLogic implementations. These objects understand the relationship between model elements and internalize the relationship as code. An example is for bundles finding their grandchildren.

* CONTRIBUTING.md updated

Once the refactoring was complete, updated the documentation to reflect the new code naming and its relationship to the architecture/design.

* docker update by Jimmie Young <jdyoung@jpl.nasa.gov>

* Add a Dockerfile for building images for AWS deployments (#151)

* Creation of AWS-specific Dockerfile, update README to indicate its purpose

* Fix typo in both Dockerfiles (extraneous ']' in CMD)

* Anonymous changes from main stomp

* Switch back to 8080 by default (#157)

* fix for 141 (#147) (#156)

* changed string searching (#158) by  Al Niessner <Al.Niessner@xxx.xxx>

Seems that wildcards is not well implemented in the later elasticsearch and opensearch implementations. The most recent documentation for either has changed significantly that if searhing for strings, then they offer a short globish like set of patterns. Moved from using wildcards to this new string and it seems to be working.

The eq operator had some problems because it was using should instead of filter for the or condition. SHould returns everything with a score value to represent the closes match from the furthest. Changed over to filter and the or condition is now behaving as expected.

* Anonymous changes from main stomp

* fixes for issue 162

* fixes for issue 162

* fix for issue 149