update API (#177)
* update API

Added all of the new endpoints in groups classes, references, and products.

Added all of the old endpoints in the group deprecrated.

* changed to fit proposal

Update swagger.yml to match the proposal plus any additions and clarification on the original ticket NASA-PDS/pds-api#202.

Re-arranged the code to fit the new API. Required some new exceptions and tweaks to the basic interfaces to support the membership concept outside of just referencing in general and that a lidvid necessarily match a particular product type.

Added new swagger transmutation layers to support the easy removal of deprecated code in the future and to support easier maintenance in the future. Left the old referencing suff only because there is a chance it may come back in the newish future to support some other use cases.

The new swagger transmutation layers still all funnel through a single processor once all of the inputs have been collected. The deprecated code mostly works itself through the new API showing that what could be done still can using the inew API. Did not want to update the standard handler to also check that desired class matches the lid(vid) so did not route those but they could also be done with a touch more code change just for that demonstration whcih is why it was not done.

* fill out verify()

Given the new endpoints of /classes/{class}/{identifier} have to verify the {identifier} is actually of the type {class}. Previous commit had a blank for this function but need it filled out to test it.

* methods need to be in controller

Seems there is a limitation in spring (maybe swagger) that the methods to be used in a @Controller must be present in the class and not inherited. This makes the code a bit more repetitive, but it works.

* adjusted regress

Had to update the endpoints to match the new API. Simply tested what already existed and did not add any new tests for new corner cases.

Had to update the number of collections and products too. Seems my latest install of data from harvest is smaller than before. However it all fits the current data in my opensearch now.