script to create latest index
Script uses harvest_date_time to know which documents to process. It allows for harvest to be running while this processing is taking place. Once it finds all documents within a given temporal range, it then checks the highest vid against what is in latest index and then update if neccessary. The script tries to minimize queries to both registry and latest indices. It also uses a scroll search to handle basically unlimited returns.