Update API default response in browsers to force application/json (#360)
* force the content-type

Part of the problem with the PDS product is that when in a browser it is better viewed using a plugin. To this end, force the return type for text/html of a PDS product or list of products to be application/json. Lovely, but if you are accessig a product directly, it is kind of what you deserve and probably desire.

Found how to set the content-type for the output. It can be overridden but required a rework of the list in WebMVCConfig. Turns out is is very order dependent. The current order seems to have everything working as desired while maybe not correctly.

It also turns out that setting the content-type is also order dependent with requesting the body. Yikes so much hidden order requirements in the bowels of spring. Anyway, marked all of the converters with comments to hopefully keep their order correct.

* Update PdsProductTextHtmlSerializer.java

put pretty print back in

* Update PdsProductTextHtmlSerializer.java

in the right place

* add text/html return as JSon for plural products responses

---------

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>
Co-authored-by: thomas loubrieu <thomas.loubrieu@jpl.nasa.gov>