existance
Updated the language to allow for finding documents that have a specific field name. If the FIELD is given, then it assumes that an exact match is being requested. If a STRINGVAL is given, then it treats it as a regular expression.

Processing regular expressions collects all of the known field names from the mapping and pulls out all matching field names. It then creates an existance check for each of the matching field names.